requirements.txt
python-telegram-bot==13.15
redis==3.2.1
python-dotenv==1.1.1
requests~=2.32.5